Common base class
