+++ tests/unit/client/auth/test_request_auth.py
@@ -22,7 +22,7 @@ def test_auth_service_account(
     """Auth can retrieve token and expiration values."""
     httpx_mock.add_callback(check_service_credentials_callback)
 
-    mocker.patch("firebolt.client.auth.service_account.time", return_value=0)
+    mocker.patch("firebolt.client.auth.request_auth_base.time", return_value=0)
     auth = ServiceAccount(mock_service_id, mock_service_secret)
     execute_generator_requests(auth.get_new_token_generator())
     assert auth.token == test_token, "invalid access token"
@@ -40,7 +40,7 @@ def test_auth_username_password(
     """Auth can retrieve token and expiration values."""
     httpx_mock.add_callback(check_credentials_callback)
 
-    mocker.patch("firebolt.client.auth.username_password.time", return_value=0)
+    mocker.patch("firebolt.client.auth.request_auth_base.time", return_value=0)
     auth = UsernamePassword(test_username, test_password)
     execute_generator_requests(auth.get_new_token_generator())
     assert auth.token == test_token, "invalid access token"
